Extract shouldBeSortedDescending into a separate file
