+++ kotest-assertions/kotest-assertions-core/src/commonMain/kotlin/io/kotest/matchers/collections/CollectionMatchers.kt
@@ -34,7 +34,6 @@ fun <T> existInOrder(predicates: List<(T) -> Boolean>): Matcher<Collection<T>?> 
 fun <T> haveSize(size: Int): Matcher<Collection<T>> = haveSizeMatcher(size)
 
 
-
 fun <T : Comparable<T>> beSorted(): Matcher<List<T>> = sorted()
 fun <T : Comparable<T>> sorted(): Matcher<List<T>> = sortedBy { it }
 
@@ -46,26 +45,6 @@ fun <T, E : Comparable<E>> sortedBy(transform: (T) -> E): Matcher<List<T>> = obj
       val elementMessage = when (failure) {
          null -> ""
          else -> ". Element ${failure.value} at index ${failure.index} was greater than element ${value[failure.index + 1]}"
-      }
-      return MatcherResult(
-         failure == null,
-         { "List ${value.print().value} should be sorted$elementMessage" },
-         { "List ${value.print().value} should not be sorted" }
-      )
-   }
-}
-
-fun <T : Comparable<T>> beSortedDescending(): Matcher<List<T>> = sortedDescending()
-fun <T : Comparable<T>> sortedDescending(): Matcher<List<T>> = sortedDescendingBy { it }
-
-fun <T, E : Comparable<E>> beSortedDescendingBy(transform: (T) -> E): Matcher<List<T>> = sortedDescendingBy(transform)
-fun <T, E : Comparable<E>> sortedDescendingBy(transform: (T) -> E): Matcher<List<T>> = object : Matcher<List<T>> {
-   override fun test(value: List<T>): MatcherResult {
-      val failure =
-         value.withIndex().firstOrNull { (i, it) -> i != value.lastIndex && transform(it) < transform(value[i + 1]) }
-      val elementMessage = when (failure) {
-         null -> ""
-         else -> ". Element ${failure.value} at index ${failure.index} was less than element ${value[failure.index + 1]}"
       }
       return MatcherResult(
          failure == null,
